requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Configuration & Settings
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pyyaml==6.0.1

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# HTTP Client for LLM APIs
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
